backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-oauth-toolkit>=3.0.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
]
